pyproject.toml
[project]
name = "rag-eval-foundations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arize-phoenix",
    "arize-phoenix-otel",
    "asyncpg",
    "ipykernel",
    "langchain",
    "langchain-cohere",
    "langchain-community",
    "langchain-core",
    "langchain-experimental",
    "langchain-openai",
    "langchain-postgres>=0.0.14",
    "openinference-instrumentation-langchain",
    "pandas",
    "psycopg2-binary",
    "python-dotenv",
    "ragas",
    "rank_bm25",
    "rapidfuzz",
    "requests",
    "sqlalchemy",
]
